restructure: about.py.
